+++ src/canvas.rs
@@ -16,7 +16,6 @@ use widgets::*;
 use crate::{
     app::{
         self,
-        data_harvester::processes::ProcessHarvest,
         layout_manager::{BottomColRow, BottomLayout, BottomWidgetType},
         App,
     },
@@ -41,7 +40,7 @@ pub struct DisplayableData {
     pub disk_data: Vec<Vec<String>>,
     pub temp_sensor_data: Vec<Vec<String>>,
     // Not the final value
-    pub process_data: HashMap<u32, ProcessHarvest>,
+    pub process_data: Vec<ConvertedProcessData>,
     // Not the final value
     pub grouped_process_data: Vec<ConvertedProcessData>,
     // What's actually displayed
@@ -386,9 +385,18 @@ impl Painter {
                 self.draw_basic_cpu(&mut f, app_state, vertical_chunks[0], 1);
                 self.draw_basic_memory(&mut f, app_state, middle_chunks[0], 2);
                 self.draw_basic_network(&mut f, app_state, middle_chunks[1], 3);
-                self.draw_basic_table_arrows(&mut f, app_state, vertical_chunks[3]);
                 if let Some(basic_table_widget_state) = &app_state.basic_table_widget_state {
                     let widget_id = basic_table_widget_state.currently_displayed_widget_id;
+
+                    if let Some(current_table) = app_state.widget_map.get(&widget_id) {
+                        self.draw_basic_table_arrows(
+                            &mut f,
+                            app_state,
+                            vertical_chunks[3],
+                            current_table,
+                        );
+                    }
+
                     match basic_table_widget_state.currently_displayed_widget_type {
                         Disk => self.draw_disk_table(
                             &mut f,