+++ src/canvas/widgets/process_table.rs
@@ -1,7 +1,7 @@
 use std::cmp::max;
 
 use crate::{
-    app::{self, App, ProcWidgetState},
+    app::{self, App},
     canvas::{
         drawing_utils::{
             get_search_start_position, get_start_position, get_variable_intrinsic_widths,
@@ -44,11 +44,11 @@ impl ProcessTableWidget for Painter {
         widget_id: u64,
     ) {
         if let Some(process_widget_state) = app_state.proc_state.widget_states.get(&widget_id) {
-            let search_width = if draw_border { 5 } else { 3 };
+            let search_height = if draw_border { 4 } else { 3 };
             if process_widget_state.is_search_enabled() {
                 let processes_chunk = Layout::default()
                     .direction(Direction::Vertical)
-                    .constraints([Constraint::Min(0), Constraint::Length(search_width)].as_ref())
+                    .constraints([Constraint::Min(0), Constraint::Length(search_height)].as_ref())
                     .split(draw_loc);
 
                 self.draw_processes_table(f, app_state, processes_chunk[0], draw_border, widget_id);
@@ -85,6 +85,9 @@ impl ProcessTableWidget for Painter {
                 // do so by hiding some elements!
                 let num_rows = max(0, i64::from(draw_loc.height) - self.table_height_offset) as u64;
                 let is_on_widget = widget_id == app_state.current_widget.widget_id;
+                let is_on_processes =
+                    is_on_widget || (widget_id + 1 == app_state.current_widget.widget_id);
+                let is_search_enabled = proc_widget_state.is_search_enabled();
 
                 let position = get_start_position(
                     num_rows,
@@ -145,7 +148,7 @@ impl ProcessTableWidget for Painter {
                 let wps = "W/s".to_string();
                 let total_read = "Read".to_string();
                 let total_write = "Write".to_string();
-                let process_state = "State".to_string();
+                // let process_state = "State".to_string();
 
                 let direction_val = if proc_widget_state.process_sorting_reverse {
                     "▼".to_string()
@@ -169,7 +172,7 @@ impl ProcessTableWidget for Painter {
                     wps,
                     total_read,
                     total_write,
-                    process_state,
+                    // process_state,
                 ];
                 let process_headers_lens: Vec<usize> = process_headers
                     .iter()
@@ -178,7 +181,7 @@ impl ProcessTableWidget for Painter {
 
                 // Calculate widths
                 let width = f64::from(draw_loc.width);
-                let width_ratios = [0.1, 0.2, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1];
+                let width_ratios = [0.1, 0.2, 0.1, 0.1, 0.1, 0.1, 0.15, 0.15];
                 let variable_intrinsic_results = get_variable_intrinsic_widths(
                     width as u16,
                     &width_ratios,
@@ -212,7 +215,7 @@ impl ProcessTableWidget for Painter {
                     String::default()
                 };
 
-                let (border_and_title_style, highlight_style) = if is_on_widget {
+                let (border_and_title_style, highlight_style) = if is_on_processes {
                     (
                         self.colours.highlighted_border_style,
                         self.colours.currently_selected_text_style,
@@ -229,7 +232,11 @@ impl ProcessTableWidget for Painter {
                         } else {
                             self.colours.widget_title_style
                         })
-                        .borders(Borders::ALL)
+                        .borders(if is_search_enabled {
+                            *TOP_LEFT_RIGHT
+                        } else {
+                            Borders::ALL
+                        })
                         .border_style(border_and_title_style)
                 } else if is_on_widget {
                     Block::default()
@@ -271,20 +278,6 @@ impl ProcessTableWidget for Painter {
         &self, f: &mut Frame<'_, B>, app_state: &mut App, draw_loc: Rect, draw_border: bool,
         widget_id: u64,
     ) {
-        fn get_prompt_text<'a>(proc_widget_state: &ProcWidgetState) -> &'a str {
-            let pid_search_text = "Search by PID (Tab for Name): ";
-            let name_search_text = "Search by Name (Tab for PID): ";
-            let grouped_search_text = "Search by Name: ";
-
-            if proc_widget_state.is_grouped {
-                grouped_search_text
-            } else if proc_widget_state.process_search_state.is_searching_with_pid {
-                pid_search_text
-            } else {
-                name_search_text
-            }
-        }
-
         fn build_query<'a>(
             is_on_widget: bool, grapheme_indices: GraphemeIndices<'a>, start_position: usize,
             cursor_position: usize, query: &str, currently_selected_text_style: tui::style::Style,
@@ -335,29 +328,16 @@ impl ProcessTableWidget for Painter {
         if let Some(proc_widget_state) =
             app_state.proc_state.widget_states.get_mut(&(widget_id - 1))
         {
-            let chosen_text = get_prompt_text(&proc_widget_state);
-
             let is_on_widget = widget_id == app_state.current_widget.widget_id;
+            let is_on_processes =
+                is_on_widget || (widget_id - 1 == app_state.current_widget.widget_id);
             let num_columns = draw_loc.width as usize;
-            let small_mode = num_columns < 70;
-            let search_title: &str = if !small_mode {
-                chosen_text
-            } else if chosen_text.is_empty() {
-                ""
-            } else if proc_widget_state.process_search_state.is_searching_with_pid
-                && !proc_widget_state.is_grouped
-            {
-                "p> "
-            } else {
-                "n> "
-            };
+            let search_title = "> ";
 
             let num_chars_for_text = search_title.len();
-
-            let mut search_text = vec![Text::styled(search_title, self.colours.table_header_style)];
-
             let cursor_position = proc_widget_state.get_cursor_position();
             let current_cursor_position = proc_widget_state.get_char_cursor_position();
+            let is_search_enabled = proc_widget_state.is_search_enabled();
 
             let start_position: usize = get_search_start_position(
                 num_columns - num_chars_for_text - 5,
@@ -373,6 +353,14 @@ impl ProcessTableWidget for Painter {
                 app_state.is_force_redraw,
             );
 
+            let mut search_text = vec![Text::styled(
+                search_title,
+                if is_on_widget {
+                    self.colours.table_header_style
+                } else {
+                    self.colours.text_style
+                },
+            )];
             let query = proc_widget_state.get_current_search_query().as_str();
             let grapheme_indices = UnicodeSegmentation::grapheme_indices(query, true);
             let query_with_cursor: Vec<Text<'_>> = build_query(
@@ -384,6 +372,8 @@ impl ProcessTableWidget for Painter {
                 self.colours.currently_selected_text_style,
                 self.colours.text_style,
             );
+
+            // TODO: [QUERY] Make text/border go red if error?
 
             // Text options shamelessly stolen from VS Code.
             let case_style = if !proc_widget_state.process_search_state.is_ignoring_case {
@@ -410,71 +400,55 @@ impl ProcessTableWidget for Painter {
                 self.colours.text_style
             };
 
-            let mut option_text = vec![];
-            let case_text = format!(
-                "{}({})",
-                if small_mode { "Case" } else { "Match Case " },
-                if self.is_mac_os { "F1" } else { "Alt+C" },
-            );
-
-            let whole_text = format!(
-                "{}({})",
-                if small_mode {
-                    "Whole"
-                } else {
-                    "Match Whole Word "
-                },
-                if self.is_mac_os { "F2" } else { "Alt+W" },
-            );
-
-            let regex_text = format!(
-                "{}({})",
-                if small_mode { "Regex" } else { "Use Regex " },
-                if self.is_mac_os { "F3" } else { "Alt+R" },
-            );
-
-            let option_row = vec![
-                Text::raw("\n\n"),
-                Text::styled(&case_text, case_style),
-                Text::raw(if small_mode { "  " } else { "     " }),
-                Text::styled(&whole_text, whole_word_style),
-                Text::raw(if small_mode { "  " } else { "     " }),
-                Text::styled(&regex_text, regex_style),
+            let option_text = vec![
+                Text::raw("\n"),
+                Text::styled(
+                    format!("Case({})", if self.is_mac_os { "F1" } else { "Alt+C" }),
+                    case_style,
+                ),
+                Text::raw("  "),
+                Text::styled(
+                    format!("Whole({})", if self.is_mac_os { "F2" } else { "Alt+W" }),
+                    whole_word_style,
+                ),
+                Text::raw("  "),
+                Text::styled(
+                    format!("Regex({})", if self.is_mac_os { "F3" } else { "Alt+R" }),
+                    regex_style,
+                ),
             ];
-            option_text.extend(option_row);
 
             search_text.extend(query_with_cursor);
+            search_text.push(Text::styled(
+                format!(
+                    "\n{}",
+                    if let Some(err) = &proc_widget_state
+                        .process_search_state
+                        .search_state
+                        .error_message
+                    {
+                        err.as_str()
+                    } else {
+                        ""
+                    }
+                ),
+                self.colours.invalid_query_style,
+            ));
             search_text.extend(option_text);
 
-            let current_border_style = if proc_widget_state
-                .process_search_state
-                .search_state
-                .is_invalid_search
-            {
-                *INVALID_REGEX_STYLE
-            } else if is_on_widget {
+            let current_border_style = if is_on_processes {
                 self.colours.highlighted_border_style
             } else {
                 self.colours.border_style
             };
 
-            let title = if draw_border {
-                const TITLE_BASE: &str = " Esc to close ";
-
-                let repeat_num = max(
-                    0,
-                    draw_loc.width as i32 - TITLE_BASE.chars().count() as i32 - 2,
-                );
-                format!("{} Esc to close ", "─".repeat(repeat_num as usize))
-            } else {
-                String::new()
-            };
-
             let process_search_block = if draw_border {
                 Block::default()
-                    .title(&title)
-                    .title_style(current_border_style)
-                    .borders(Borders::ALL)
+                    .borders(if is_search_enabled {
+                        *BOTTOM_LEFT_RIGHT
+                    } else {
+                        Borders::ALL
+                    })
                     .border_style(current_border_style)
             } else if is_on_widget {
                 Block::default()