+++ src/utils/error.rs
@@ -1,4 +1,4 @@
-use std::result;
+use std::{borrow::Cow, result};
 
 /// A type alias for handling errors related to Bottom.
 pub type Result<T> = result::Result<T, BottomError>;
@@ -22,6 +22,8 @@ pub enum BottomError {
     ConfigError(String),
     /// An error to represent errors with converting between data types.
     ConversionError(String),
+    /// An error to represent errors with querying.
+    QueryError(Cow<'static, str>),
 }
 
 impl std::fmt::Display for BottomError {
@@ -47,6 +49,7 @@ impl std::fmt::Display for BottomError {
             BottomError::ConversionError(ref message) => {
                 write!(f, "unable to convert: {}", message)
             }
+            BottomError::QueryError(ref message) => write!(f, "{}", message),
         }
     }
 }
@@ -96,5 +99,11 @@ impl From<fern::InitError> for BottomError {
 impl From<std::str::Utf8Error> for BottomError {
     fn from(err: std::str::Utf8Error) -> Self {
         BottomError::ConversionError(err.to_string())
+    }
+}
+
+impl From<regex::Error> for BottomError {
+    fn from(err: regex::Error) -> Self {
+        BottomError::QueryError(err.to_string().into())
     }
 }